SG2044: Rename PeilessSec to SecMain
Rename SEC module from PeilessSec to SecMain and update all
SG2044 platform .dsc/.fdf references accordingly.
