+++ Platform/Sophgo/SG2044Pkg/SD3-10-LB/SD3-10-LB.dsc
@@ -752,7 +752,7 @@
   #
   # SEC Phase modules
   #
-  Silicon/Sophgo/Core/Sec/PeilessSec.inf  {
+  Silicon/Sophgo/Core/Sec/SecMain.inf  {
     <LibraryClasses>
       ExtractGuidedSectionLib|EmbeddedPkg/Library/PrePiExtractGuidedSectionLib/PrePiExtractGuidedSectionLib.inf
       LzmaDecompressLib|MdeModulePkg/Library/LzmaCustomDecompressLib/LzmaCustomDecompressLib.inf
