+++ Platform/Sophgo/SG2044Pkg/SRA3-40-LB/SRA3-40-LB.dsc
@@ -761,7 +761,7 @@
   #
   # SEC Phase modules
   #
-  Silicon/Sophgo/Core/Sec/PeilessSec.inf  {
+  Silicon/Sophgo/Core/Sec/SecMain.inf  {
     <LibraryClasses>
       ExtractGuidedSectionLib|EmbeddedPkg/Library/PrePiExtractGuidedSectionLib/PrePiExtractGuidedSectionLib.inf
       LzmaDecompressLib|MdeModulePkg/Library/LzmaCustomDecompressLib/LzmaCustomDecompressLib.inf
